Move icons under icons/berthReservations
