+++ src/common/icons/berthReservations/IconArrowLeft.tsx
