+++ src/common/icons/berthReservations/IconArrowRight.tsx
