+++ src/common/icons/berthReservations/IconDollyEmpty.tsx
