+++ src/common/icons/berthReservations/IconInvoice.tsx
