+++ src/common/icons/berthReservations/IconNoBoat.tsx
