+++ src/common/icons/berthReservations/IconRegisteredBoat.tsx
