+++ src/common/icons/berthReservations/IconStreetLight.tsx
