+++ src/common/icons/berthReservations/IconTrash.tsx
