+++ src/common/icons/berthReservations/IconWinter.tsx
