+++ src/common/icons/index.ts
@@ -1,28 +1,28 @@
 export type { IconProps } from './Icon.interface';
 
-export { default as IconAccessibility } from './IconAccessibility';
-export { default as IconAngleRight } from './IconAngleRight';
-export { default as IconApplications } from './IconApplications';
-export { default as IconArrowLeft } from './IconArrowLeft';
-export { default as IconArrowRight } from './IconArrowRight';
-export { default as IconBoat } from './IconBoat';
-export { default as IconBusiness } from './IconBusiness';
-export { default as IconContracts } from './IconContracts';
-export { default as IconDirections } from './IconDirections';
-export { default as IconDivided } from './IconDivided';
-export { default as IconDollyEmpty } from './IconDollyEmpty';
-export { default as IconFence } from './IconFence';
-export { default as IconInvoice } from './IconInvoice';
-export { default as IconNoBoat } from './IconNoBoat';
-export { default as IconPerson } from './IconPerson';
-export { default as IconPlug } from './IconPlug';
-export { default as IconPole } from './IconPole';
-export { default as IconRead } from './IconRead';
-export { default as IconRegisteredBoat } from './IconRegisteredBoat';
-export { default as IconStreetLight } from './IconStreetLight';
-export { default as IconTimes } from './IconTimes';
-export { default as IconTrash } from './IconTrash';
-export { default as IconTrestle } from './IconTrestle';
-export { default as IconUnregisteredBoat } from './IconUnregisteredBoat';
-export { default as IconWaterTap } from './IconWaterTap';
-export { default as IconWinter } from './IconWinter';
+export { default as IconAccessibility } from './berthReservations/IconAccessibility';
+export { default as IconAngleRight } from './berthReservations/IconAngleRight';
+export { default as IconApplications } from './berthReservations/IconApplications';
+export { default as IconArrowLeft } from './berthReservations/IconArrowLeft';
+export { default as IconArrowRight } from './berthReservations/IconArrowRight';
+export { default as IconBoat } from './berthReservations/IconBoat';
+export { default as IconBusiness } from './berthReservations/IconBusiness';
+export { default as IconContracts } from './berthReservations/IconContracts';
+export { default as IconDirections } from './berthReservations/IconDirections';
+export { default as IconDivided } from './berthReservations/IconDivided';
+export { default as IconDollyEmpty } from './berthReservations/IconDollyEmpty';
+export { default as IconFence } from './berthReservations/IconFence';
+export { default as IconInvoice } from './berthReservations/IconInvoice';
+export { default as IconNoBoat } from './berthReservations/IconNoBoat';
+export { default as IconPerson } from './berthReservations/IconPerson';
+export { default as IconPlug } from './berthReservations/IconPlug';
+export { default as IconPole } from './berthReservations/IconPole';
+export { default as IconRead } from './berthReservations/IconRead';
+export { default as IconRegisteredBoat } from './berthReservations/IconRegisteredBoat';
+export { default as IconStreetLight } from './berthReservations/IconStreetLight';
+export { default as IconTimes } from './berthReservations/IconTimes';
+export { default as IconTrash } from './berthReservations/IconTrash';
+export { default as IconTrestle } from './berthReservations/IconTrestle';
+export { default as IconUnregisteredBoat } from './berthReservations/IconUnregisteredBoat';
+export { default as IconWaterTap } from './berthReservations/IconWaterTap';
+export { default as IconWinter } from './berthReservations/IconWinter';
